Changed: SettingKeys and methods in DevOopsGitReleasePlugin.autoImport to traits
